Refactored the program using a custom view
Instead of adding a canvas to a bitmap, and attaching the bitmap to an ImageView. I refactored the program by deleting the bitmap, ImageView components, and added a custom view TreeCustomView, to simplify the logic of the program. But still need to figure out how to update the custom view once a button is clicked.